+++ src/client/provider.test.ts
@@ -1,5 +1,12 @@
 import { Provider, ProviderConfiguration } from './provider';
 
+describe('Provider Configuration', () => {
+  it('should cache key correctly', async () => {
+    const mockProviderConfiguration = new ProviderConfiguration('test', []);
+    expect(mockProviderConfiguration.cacheKey).toEqual('{"provider":"test"}');
+  });
+});
+
 describe('Provider', () => {
   it('configures provider correctly', async () => {
     const mockProviderConfiguration = new ProviderConfiguration('test', []);
@@ -38,31 +45,8 @@ describe('Provider', () => {
     );
     const mockProvider = new Provider(mockProviderConfiguration);
 
-    await expect(
-      mockProvider.configure({
-        security: [
-          {
-            username: 'second',
-            password: 'seconds',
-            id: 'second-id',
-          },
-        ],
-      })
-    ).resolves.toEqual(
-      new Provider(
-        new ProviderConfiguration('test', [
-          {
-            id: 'first-id',
-            username: 'digest-user',
-            password: 'digest-password',
-          },
-          {
-            username: 'second',
-            password: 'seconds',
-            id: 'second-id',
-          },
-        ])
-      )
+    await expect(mockProvider.configure()).resolves.toEqual(
+      new Provider(new ProviderConfiguration('test', mockSecurity))
     );
   });
 });