+++ src/client/usecase.ts
@@ -82,7 +82,9 @@ class UseCaseBase implements Interceptable {
     this.configureHookContext();
   }
 
-  private async bind(options?: PerformOptions): Promise<void> {
+  private async bind(options?: {
+    provider?: string | Provider | undefined;
+  }): Promise<void> {
     const hookRouter =
       this.events.hookContext[`${this.profileConfiguration.id}/${this.name}`]
         .router;
@@ -133,15 +135,7 @@ class UseCaseBase implements Interceptable {
     currentProvider: string | undefined,
     options?: PerformOptions
   ) {
-    const providerConfig = await this.getProviderConfiguration(
-      options?.provider ?? currentProvider
-    );
-    // If we have security values pass them directly to perform
-    if (options?.security) {
-      return new ProviderConfiguration(providerConfig.name, options.security);
-    }
-
-    return providerConfig;
+    return this.getProviderConfiguration(options?.provider ?? currentProvider);
   }
 
   private async getProviderConfiguration(
@@ -173,7 +167,8 @@ class UseCaseBase implements Interceptable {
     TOutput = unknown
   >(
     input?: TInput,
-    parameters?: Record<string, string>
+    parameters?: Record<string, string>,
+    security?: SecurityValues[]
   ): Promise<Result<TOutput, PerformError>> {
     if (this.boundProfileProvider === undefined) {
       throw new UnexpectedError(
@@ -185,7 +180,8 @@ class UseCaseBase implements Interceptable {
     return this.boundProfileProvider.perform<TInput, TOutput>(
       this.name,
       input,
-      parameters
+      parameters,
+      security
     );
   }
 
@@ -204,7 +200,11 @@ class UseCaseBase implements Interceptable {
 
     this.log?.('Bound provider: %O', this.boundProfileProvider);
 
-    return this.performBoundUsecase(input, options?.parameters);
+    return this.performBoundUsecase(
+      input,
+      options?.parameters,
+      options?.security
+    );
   }
 
   private checkWarnFailoverMisconfiguration() {